projeto_e-Commerce-logistics/pyproject.toml
[project]
name = "projeto-e-commerce-logistics"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dbt-core>=1.11.6",
    "dbt-duckdb>=1.10.1",
    "delta-spark>=4.1.0",
    "psycopg2-binary>=2.9.11",
    "pyspark>=4.1.1",
    "python-dotenv>=1.2.1",
]
